legal_simplifier/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.1
python-dotenv==1.0.1
langchain-groq==0.1.3
langchain==0.2.16
easyocr==1.7.0
reportlab==4.0.4
PyPDF2==3.0.1